VideoProcessingService/requirements.txt
typing>=3.5.2
opencv-contrib-python>=3.4.0
imutils>=0.4.5
kafka-python>=1.3.5
